+++ packages/core/admin/admin/src/layouts/UnauthenticatedLayout/LocaleToggle/index.js
@@ -7,7 +7,7 @@
 import React from 'react';
 import { useIntl } from 'react-intl';
 import { SimpleMenu, MenuItem } from '@strapi/parts';
-import useLocalesProvider from '../LocalesProvider/useLocalesProvider';
+import useLocalesProvider from '../../../components/LocalesProvider/useLocalesProvider';
 
 const LocaleToggle = () => {
   const { changeLocale, localeNames } = useLocalesProvider();
@@ -16,7 +16,9 @@ const LocaleToggle = () => {
   return (
     <SimpleMenu label={localeNames[locale]}>
       {Object.keys(localeNames).map(lang => (
-        <MenuItem onClick={() => changeLocale(lang)} key={lang}>{localeNames[lang]}</MenuItem>
+        <MenuItem onClick={() => changeLocale(lang)} key={lang}>
+          {localeNames[lang]}
+        </MenuItem>
       ))}
     </SimpleMenu>
   );
