+++ packages/core/admin/admin/src/pages/Admin/Onboarding/StaticLinks/StyledLink.js
